kernel: clear advisory clang-tidy bug-class findings + fix ACPI direct-map assumption
Two changes, combined per request:

1. clang-tidy bug-class cleanup (the ~45 non-style advisory findings;
   the 894 readability-braces / loop-convert / using-namespace style
   findings are intentionally left — they are advisory-only noise):

   - bugprone-argument-comment (12): net_smoke.cpp / tls.cpp call-site
     /*name=*/ comments updated to the real parameter names.
   - readability-inconsistent-declaration-parameter-name (5): mlme.h /
     eapol.h / fourway.h declarations aligned to their definitions.
   - bugprone-branch-clone (3): drsh_desktop.cpp removed a dead,
     immediately-overwritten buggy `(p[3]!=0)?false:false` assignment
     (real latent bug); tcp.cpp / beacon.cpp merged genuinely-shared
     branch bodies with no behaviour change (beacon.cpp also drops the
     now-redundant has_psk to avoid -Werror=unused).
   - bugprone-implicit-widening-of-multiplication-result (1 site):
     drsh_desktop.cpp framebuffer offset math widened to u64 before
     the multiplies.
   - bugprone-macro-parentheses (1): result.h RESULT_TRY_ASSIGN — decl
     is a declarator, not an expression; documented NOLINT.
   - bugprone-dynamic-static-initializers (16): tcp_internal.h /
     shell_internal.h statics made `constinit` at both the extern
     decls and definitions (this surfaced + fixed real non-constant
     init of Tcb/Stats, which now use explicit value-init). The check
     still flags extern-in-header so a documented NOLINT closes it —
     the freestanding hazard itself is eliminated by constinit.
   - performance-enum-size (6): VfsBackend->u8, KObjectType->u16,
     ErrorCode->u8 (internal, no wire format); Cap / FwPackageFamily /
     FwPackageFlags kept wide with documented NOLINT (ABI-adjacent /
     serialized field widths — narrowing them would be wrong).

2. ACPI: stop assuming every ACPI table lives below the 1 GiB direct
   map. PhysToHeader() now routes through a cached AcpiMapPhys() that
   uses PhysToVirt when in range and MapMmio otherwise. QEMU/OVMF
   (tables low) is unchanged; VirtualBox parks the XSDT near the top of
   2 GiB RAM, which previously hard-panicked "PhysToVirt called outside
   direct map" at boot. Resolves the long-standing acpi.h TODO. DSDT /
   SSDT scanners routed through the same helper; acpi.h scope note
   updated.

Verified: x86_64-release + x86_64-debug build zero-warning/zero-error,
ctest green, QEMU bringup smoke reaches bringup-complete with no fatal
sentinels. clang-tidy bug-class findings: 0 remaining. The MapMmio ACPI
path only triggers on VirtualBox-style high table placement (not QEMU),
so it is validated on real VirtualBox boot.

+++ kernel/ipc/kobject.h
@@ -56,7 +56,7 @@ namespace duetos::ipc
 /// Type tag stored in every KObject. The numeric values are stable
 /// — used by `inspect` / panic dumps. New types append at the end;
 /// never re-use a retired tag.
-enum class KObjectType : u32
+enum class KObjectType : u16
 {
     Invalid = 0,
     Mutex = 1,
